Split ElevenLabs STT runtime helpers
